voice-server/pyproject.toml
[project]
name = "pravik-voice-server"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pipecat-ai[silero,openai,websocket]>=0.0.60",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "supabase>=2.13.0",
    "httpx>=0.28.0",
    "twilio>=9.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
